.tools/test/stacks/admin/python/requirements.txt
boto3
aws-cdk-lib>=2.0.0
constructs>=10.0.0
types-boto3
types-setuptools
pyyaml
random2